Fixes Docker build issues
This PR resolves issues relating to the Docker registry build process.

The base image that we previously used is no longer available, the Dockerfile
has been updated to reflect that change.  Also since Debain Jessie is used instead of
whezzy we had to use unstable to provide fuseiso packages.

Removed the `git clone` for the project and replaced with a Dockerfile COPY command.
This change will make it easier for user forks to create containers based on their current branch.

Updated the Gemfile and Gemfile.lock based on Ruby version changes from 2.2.1 to 2.2.3

Testing will follow.