Adding docker-compose to automate the containers for Hanlon
The Hanlon container requires three other services to run.
The database, typically mongo, tftp and DHCP.  We have created a
container for tftp and use mongo's already available container
on Docker registry.  DHCPd is just easier at this point to run on
the host or an externally so that is not included.

This docker-compose addition will orchestrate the starting of the
required services for Hanlon.

We also added a `.dockerignore` so the container build process
removes those unneeded files.